Real constructors from signed numbers accessible for Kotlin+Java/JVM, and invoke-emulated constructors from unsigned numbers only for Kotlin (because Java doesn't know these).

+++ library/src/commonMain/kotlin/DecimalExtensions.kt
@@ -30,13 +30,13 @@ public fun Float.toDecimal():Decimal = Decimal(this)
 public fun Double.toDecimal():Decimal = Decimal(this)
 
 public fun Byte.toDecimal():Decimal = Decimal(this)
-public fun UByte.toDecimal():Decimal = Decimal(this)
+public fun UByte.toDecimal():Decimal = Decimal(this.toByte())
 public fun Short.toDecimal():Decimal = Decimal(this)
-public fun UShort.toDecimal():Decimal = Decimal(this)
+public fun UShort.toDecimal():Decimal = Decimal(this.toShort())
 public fun Int.toDecimal():Decimal = Decimal(this)
-public fun UInt.toDecimal():Decimal = Decimal(this)
+public fun UInt.toDecimal():Decimal = Decimal(this.toInt())
 public fun Long.toDecimal():Decimal = Decimal(this)
-public fun ULong.toDecimal():Decimal = Decimal(this)
+public fun ULong.toDecimal():Decimal = Decimal(this.toLong())
 
 
 /****************************  Convenience Suffix .Dc  ***************************/
